requirements.txt
opencc-python-reimplemented==0.1.5
